chore/#16: 패키징 수정
