+++ feature/src/main/java/com/sopt/anshim/feature/addbook/component/FilledTextField.kt
@@ -1,6 +1,7 @@
-package com.sopt.anshim.feature.ui.addBook.component
+package com.sopt.anshim.feature.addbook.component
 
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -10,26 +11,23 @@ import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun NumericInputField(
-    value: String,
+fun FilledTextField(
+    text : String,
     onValueChange: (String) -> Unit,
-    label: String
+    label : String,
 ) {
     OutlinedTextField(
-        value = value,
-        onValueChange = {
-            if (it.all { char -> char.isDigit() } || it.isBlank()) {
-                onValueChange(it)
-            }
-        },
-        label = { Text(label) },
+        value = text,
+        onValueChange = onValueChange,
+        label = { Text(label, fontSize = 16.sp, color = Color.Gray) },
         singleLine = true,
-        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
+        keyboardOptions = KeyboardOptions.Default.copy(imeAction = ImeAction.Next),
         shape = RoundedCornerShape(8.dp),
         colors = TextFieldDefaults.outlinedTextFieldColors(
             focusedBorderColor = Color.Blue,
@@ -41,6 +39,7 @@ fun NumericInputField(
             unfocusedPlaceholderColor = Color.Gray
         ),
         modifier = Modifier
+            .padding(bottom = 8.dp)
             .fillMaxWidth()
     )
 }
