+++ feature/src/main/java/com/sopt/anshim/feature/addbook/component/NumericInputField.kt
@@ -1,7 +1,6 @@
-package com.sopt.anshim.feature.ui.addBook.component
+package com.sopt.anshim.feature.addbook.component
 
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -11,23 +10,26 @@ import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun FilledTextField(
-    text : String,
+fun NumericInputField(
+    value: String,
     onValueChange: (String) -> Unit,
-    label : String,
+    label: String
 ) {
     OutlinedTextField(
-        value = text,
-        onValueChange = onValueChange,
-        label = { Text(label, fontSize = 16.sp, color = Color.Gray) },
+        value = value,
+        onValueChange = {
+            if (it.all { char -> char.isDigit() } || it.isBlank()) {
+                onValueChange(it)
+            }
+        },
+        label = { Text(label) },
         singleLine = true,
-        keyboardOptions = KeyboardOptions.Default.copy(imeAction = ImeAction.Next),
+        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
         shape = RoundedCornerShape(8.dp),
         colors = TextFieldDefaults.outlinedTextFieldColors(
             focusedBorderColor = Color.Blue,
@@ -39,7 +41,6 @@ fun FilledTextField(
             unfocusedPlaceholderColor = Color.Gray
         ),
         modifier = Modifier
-            .padding(bottom = 8.dp)
             .fillMaxWidth()
     )
 }
